requirements.txt
pdf2image==1.17.0
Pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
tqdm==4.66.1
psutil==5.9.6
argparse
xinference[all]
transformers>=4.37.0
torch>=2.1.0
torchvision>=0.16.0
qwen-vl-utils>=0.0.8
accelerate>=0.20.0
sentencepiece>=0.1.98
tiktoken>=0.5.0